refactor: context 분리
